Revert "fix(ark-runtime): support fractional task duration"
This reverts commit 9e5b6ab9b7fe7201cbbc1f21b1935166fe143a9a.

+++ volcenginesdkarkruntime/types/content_generation/content_generation_task.py
@@ -108,7 +108,7 @@ class ContentGenerationTask(BaseModel):
     generate_audio: bool
     """Whether to generate audio."""
 
-    duration: float
+    duration: int
     """The duration of the generated content in seconds."""
 
     ratio: str
